lesson_26/classwork_26_01/requirements.txt
pytest==3.6.3
selenium==3.14.0
pytest-html==1.19.0
webdriver-manager==1.7
allure-pytest==2.19.0
pytest-rerunfailures==4.1
pytest-testconfig==0.1.0
pyvirtualdisplay

